Part_03/requirements.txt
Flask==1.1.1
google-cloud-bigquery==1.22.0
gunicorn